add isOnline
