+++ frontend/package.json
@@ -35,6 +35,8 @@
     "react-native-modal-datetime-picker": "^17.1.0",
     "react-native-safe-area-context": "4.10.1",
     "react-native-screens": "3.31.1",
+    "react-native-svg": "^15.3.0",
+    "react-native-svg-transformer": "^1.4.0",
     "react-native-vector-icons": "^10.1.0",
     "react-native-web": "~0.19.10"
   },
